requirements.txt
numpy>=2.0
langchain>=0.3
langchain-core>=0.3
langchain-openai>=1.0
langgraph>=0.4
python-dotenv>=1.0
langfuse>=2.0
